requirements.txt
fastapi
python-dotenv==1.0.0
requests==2.31.0
starlette==0.27.0
uvicorn==0.22.0
boto3==1.36.8 
botocore==1.36.8
celery[redis]==5.4.0 
PyMuPDF==1.25.2
openai==1.60.2
pillow==11.1.0 
python-multipart==0.0.20
google-genai==1.2.0 